Create README.md
Overview
Welcome to the Amazon Clone project, a meticulously crafted replica of the world-renowned e-commerce platform, Amazon. This repository houses the source code for a web application built entirely with HTML and CSS.My goal was to create a realistic online shopping experience that showcases not only the visual prowess of our design but also the attention to detail in replicating Amazon's functionality.